What type of PR is this?  sync slot What this PR does / why we need it:
Which issue(s) this PR fixes:

Special notes for your reviewer:

Does this PR introduce a user-facing change?:

What type of PR is this?
 hashslot: leader sync from session; follower sync from leader
What this PR does / why we need it:

Which issue(s) this PR fixes:

Special notes for your reviewer:

Does this PR introduce a user-facing change?:

What type of PR is this?
 sync slot
What this PR does / why we need it:

Which issue(s) this PR fixes:

Special notes for your reviewer:

Does this PR introduce a user-facing change?: